admin page components in separate folders
